# code/claims-service/src/main/resources/openapi/claims-api.yml
openapi: 3.0.3
info:
  title: Claims Service API
  description: |
    Internal Claims Service API for the Demo Insurance Claims Application.
    This is the single source of truth for the BFF ↔ Claims Service contract.
    All endpoints require OAuth 2.0 bearer token authentication.
  version: 1.0.0
  contact:
    name: Demo App Team
servers:
  - url: http://localhost:8081
    description: Local development server
...
paths:
...
  /api/users:
    get:
      tags:
        - Users
      operationId: listAllUsers
      summary: List all users (Admin only)
      description: Returns a list of all users in the system. Requires ADMIN role.
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
              example:
                - userId: "550e8400-e29b-41d4-a716-446655440000"
                  email: admin@example.com
                  name: Admin User
                  role: ADMIN
                  createdAt: "2026-02-17T10:30:00.000Z"
                - userId: "660e8400-e29b-41d4-a716-446655440001"
                  email: jane.doe@example.com
                  name: Jane Doe
                  role: CLAIMANT
                  createdAt: "2026-02-17T11:00:00.000Z"
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Access denied - requires ADMIN role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    post:
      tags:
        - Users
      operationId: createUser
      summary: Create a new user
      description: Creates a new user in the Claims Service.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
            example:
              email: jane.doe@example.com
              name: Jane Doe
              role: CLAIMANT
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              example:
                userId: "550e8400-e29b-41d4-a716-446655440000"
                email: jane.doe@example.com
                name: Jane Doe
                role: CLAIMANT
                createdAt: "2026-02-17T10:30:00.000Z"
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: "https://api.demo-app.com/errors/validation-error"
                title: Validation Error
                status: 400
                detail: "Email format is invalid"
                instance: "/api/users"
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: "https://api.demo-app.com/errors/unauthorized"
                title: Unauthorized
                status: 401
                detail: "Bearer token is missing or invalid"
                instance: "/api/users"
        '409':
          description: User with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: "https://api.demo-app.com/errors/duplicate-entity"
                title: Conflict
                status: 409
                detail: "A user with this email already exists"
                instance: "/api/users"
  /api/users/{userId}:
    get:
      tags:
        - Users
      operationId: getUser
      summary: Get user details
      description: Returns the details of a user by their ID.
      parameters:
        - name: userId
          in: path
          required: true
          description: The unique identifier of the user
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              example:
                userId: "550e8400-e29b-41d4-a716-446655440000"
                email: jane.doe@example.com
                name: Jane Doe
                role: CLAIMANT
                createdAt: "2026-02-17T10:30:00.000Z"
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: "https://api.demo-app.com/errors/not-found"
                title: Not Found
                status: 404
                detail: "User not found"
                instance: "/api/users/550e8400-e29b-41d4-a716-446655440000"

  # =========================================================================
  # Claim Management Endpoints
  # =========================================================================
...
components:
  schemas:
...
    CreateUserRequest:
      type: object
      required:
        - userId
        - email
        - name
        - role
      properties:
        userId:
          type: string
          format: uuid
          description: User ID from Keycloak (must match Keycloak user ID)
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          description: User's email address
          example: jane.doe@example.com
        name:
          type: string
          minLength: 2
          maxLength: 100
          description: User's full name
          example: Jane Doe
        role:
          $ref: '#/components/schemas/UserRole'
    UserResponse:
      type: object
      required:
        - userId
        - email
        - name
        - role
        - createdAt
      properties:
        userId:
          type: string
          format: uuid
          description: Unique user identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          description: User's email address
          example: jane.doe@example.com
        name:
          type: string
          description: User's full name
          example: Jane Doe
        role:
          $ref: '#/components/schemas/UserRole'
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp (ISO 8601)
          example: "2026-02-17T10:30:00.000Z"
    UserRole:
      type: string
      enum:
        - CLAIMANT
        - ADMIN
      description: User role for authorization

    # =========================================================================
    # Claim Schemas
    # =========================================================================
...
    ProblemDetail:
      type: object
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type
          example: "https://api.demo-app.com/errors/validation-error"
        title:
          type: string
          description: A short, human-readable summary of the problem type
          example: Validation Error
        status:
          type: integer
          description: The HTTP status code
          example: 400
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence
          example: "User ID is required"
        instance:
          type: string
          description: A URI reference that identifies the specific occurrence
          example: "/api/claims"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth 2.0 bearer token from Keycloak
security:
  - bearerAuth: []